test: refactor utils for multichain components
