+++ app/component-library/components-temp/MultichainAccounts/AccountCell/AccountCell.test.tsx
@@ -1,19 +1,14 @@
 import React from 'react';
 import { AccountGroupObject } from '@metamask/account-tree-controller';
-import { AccountGroupType } from '@metamask/account-api';
 import { AccountCell } from './AccountCell';
 import renderWithProvider from '../../../../util/test/renderWithProvider';
+import { createMockAccountGroup } from '../test-utils';
 
-const mockAccountGroup: AccountGroupObject = {
-  type: AccountGroupType.SingleAccount,
-  id: 'keyring:test-group/ethereum' as const,
-  accounts: ['account-1'],
-  metadata: {
-    name: 'Test Account Group',
-    pinned: false,
-    hidden: false,
-  },
-};
+const mockAccountGroup = createMockAccountGroup(
+  'keyring:test-group/ethereum',
+  'Test Account Group',
+  ['account-1'],
+);
 
 const renderAccountCell = (
   props: {
